WIP: Make the quick access filter more flexible
The format right now is:

<res1[,res2,res3]>\s<name1[,name2,name3]>

where 'name' can be:

- `*`
- `<name>`
- `<namemspace>/*`
- `<namespace>/<name>`

The idea is to keep this (simple) format but allow multiple ;-separated filters:

<filter1>[;<filter2>;<filter3>]

This PR has a PoC, but I'm not sure it's the most efficient implementation,
and the UI performance should remain fast enough. Plus, adding some tests would
be great.